requirements.txt
Django
djangorestframework
psycopg2-binary
pytest-django
factory_boy
mock
coverage
pre-commit
black
flake8
isort
interrogate
django-bootstrap4
djangorestframework-jwt
parameterized
pytest-order
django-filter
django-cors-headers
